# Web/server/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "EGH455 Assignment",
		"version": "1.0.0",
		"description": "Documentation for the EGH455 Web APIs"
	},
	"servers": [{ "url": "http://192.168.1.13:5000" }, { "url": "http://localhost:5000" }],

	"tags": [
		{
			"name": "Sensor",
			"description": "Endpoint for sending/retrieving sensor data"
		},
		{
			"name": "Image",
			"description": "Endpoint for sending/retrieving image data"
		},
		{
			"name": "Target",
			"description": "Endpoint for sending/retrieving target data"
		}
	],
	"paths": {
		"/sensor": {
			"get": {
				"summary": "Get sensor data from the database",
				"tags": ["Sensor"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SensorData"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Store sensor data, and send it to the web interface",
				"tags": ["Sensor"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SensorData"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/image": {
			"post": {
				"summary": "Send image data to the web interface",
				"tags": ["Image"],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/ImageData"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/target": {
			"get": {
				"summary": "Get target data from the database",
				"tags": ["Target"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/TargetData"
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"SensorData": {
				"type": "object",
				"properties": {
					"temperature": {
						"type": "number"
					},
					"pressure": {
						"type": "number"
					},
					"humidity": {
						"type": "number"
					},
					"light": {
						"type": "number"
					},
					"oxidised": {
						"type": "number"
					},
					"reduced": {
						"type": "number"
					},
					"nh3": {
						"type": "number"
					}
				}
			},
			"ImageData": {
				"type": "object",
				"properties": {
					"file": {
						"type": "string",
						"format": "binary"
					},
					"personDetected": {
						"type": "string",
						"enum": ["True", "False"]
					},
					"backpackDetected": {
						"type": "string",
						"enum": ["True", "False"]
					},
					"markerDetected": {
						"type": "string",
						"enum": ["True", "False"]
					}
				},
				"required": ["file"]
			},
			"TargetData": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string"
					},
					"date": {
						"type": "string"
					}
				}
			}
		}
	}
}